frontend/requirements.txt
altair==4.2.2
anyio==3.6.2
argon2==0.1.10
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
extra-streamlit-components==0.1.56
Faker==18.6.2
favicon==0.7.0
fonttools==4.39.3
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
htbuilder==0.6.1
idna==3.4
importlib-metadata==6.6.0
isort==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
loguru==0.7.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
markdownlit==0.0.7
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
more-itertools==9.1.0
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
protobuf==3.20.3
pyarrow==12.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.6.0
pymdown-extensions==9.11
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.30.0
rich==13.3.5
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
st-annotated-text==4.0.0
starlette==0.26.1
streamlit==1.22.0
streamlit-authenticator==0.2.2
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-cookies-manager==0.2.0
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.3
streamlit-keyup==0.2.0
streamlit-login-auth-ui==0.2.0
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.1
trycourier==4.4.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.2
uvicorn==0.22.0
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0
plotly==5.15.0
loguru
av==10.0.0